+++ CHANGELOG.rst
@@ -70,6 +70,8 @@ odd bug fix may not get mentioned here at all.
     provide a list of completions for attributes that require dynamic generation
     of the possible completions beyond what is provided by a schema.
 
+  - Feature: support for automatic attribute hiding.
+
   - GUI Fix: When the user would use the down arrow to navigate the options of a
     completion menu, the focus would be lost from the document and would not be
     regained when the user closes the completion menu. This made further typing